Fixed randomTx implementation by adding the Tx timestamp to the packet

+++ src/deck/drivers/interface/TwrSwarm/TwrSwarmDebug.h
@@ -13,8 +13,8 @@ typedef struct {
   uint32_t remoteReply;
   uint32_t localReply;
   uint32_t localRound;
-  float clockCorrection;
-  float clockCorrectionCandidate;
+  uint32_t clockCorrection;
+  uint32_t clockCorrectionCandidate;
 
   uint32_t tof;
